Add tests for check_dependencies.py
